Migrated the navigations to a BEVM model
The reasons of the change:
I changed it because wasn't totally clear how the code worked. We had "three" navigations:
    1.- The block where the other navigations were supposed to inherited their basic styles.
    2.- The main navigation used in the major layout (it was a completely different component, all the styles that were inherited from the block were overwritten, thus it wasn't clear how it worked).
    3.- The navigation used downside in the nesting(1) page (all its styles were inherited from the block, but there were some things declared in the nesting's partial that made unclear if it had been reused or if it was a different component).

How it works now:
What I did was to create a different component for each navigation instead of creating a block that "inherited" their styles. Now exists the "nav__list-main" (for the major layout) and the "nav__list" (as a reusable component), this way each component has a single responsability (this means that it's styles aren't related to it's inheritance anymore which make the use of the component more clear).

Furthermore, now, instead of adding the "ly-basics" class in the HTML to each component we create, the major components already have their basic styles declared with the @extend at-rule in SASS (this make the task of creating new pages with the same major components a less frustrated without generating a lot of code). The same was added to the minor components, but in this case, if you want to create a new component that haven't beend declared in the SASS already, you'll have to extend the basic styles manually (once only, then you can reuse that component as many times as you want).

Additional information:
I added a new helper for declarating a padding of 0 with the -u prefix (in order to follow the methodology that Envato uses in its marketplace which is BEVM plus helpers with the "-u" prefix, and states with the "-is" prefix).